Remove test without index from MySQL
